#18: rename files and update file path
