+++ waste-lookup/src/components/FavoriteLists/FavoriteList.css
