+++ waste-lookup/src/components/FavoriteLists/FavoriteList.js
