+++ waste-lookup/src/components/Header/Header.css
