+++ waste-lookup/src/components/SearchBar/SearchBar.css
