+++ waste-lookup/src/components/SearchBar/SearchBar.js
