+++ waste-lookup/src/components/SearchResults/SearchResult.css
