+++ waste-lookup/src/components/SearchResults/SearchResult.js
